requirements.txt
langchain==0.1.14
langchain_community==0.0.31
langchain_openai==0.1.1
openai==1.16.1
pandas==2.2.1
python-dotenv==1.0.1
streamlit==1.31.1
chromadb